requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-multipart==0.0.6
psycopg2-binary==2.9.9